requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
tavily-python==0.3.3
pydantic==2.7.1
pydantic-settings==2.3.0
slowapi==0.1.9
python-multipart==0.0.9
aiofiles==23.2.1
